+++ drivers/aps6404/aps6404.cmake
@@ -3,6 +3,7 @@ add_library(${DRIVER_NAME} INTERFACE)
 
 target_sources(${DRIVER_NAME} INTERFACE
   ${CMAKE_CURRENT_LIST_DIR}/aps6404.cpp
+  ${CMAKE_CURRENT_LIST_DIR}/aps6404_perf_critical.cpp
 )
 
 pico_generate_pio_header(${DRIVER_NAME} ${CMAKE_CURRENT_LIST_DIR}/aps6404.pio)
