+++ drivers/aps6404/aps6404.cpp
@@ -18,6 +18,53 @@ extern "C" {
 }
 #endif
 
+namespace {
+    void aps6404_reset_program_init(PIO pio, uint sm, uint offset, uint csn, uint mosi) {
+        uint miso = mosi + 1;
+        pio_gpio_init(pio, csn);
+        pio_gpio_init(pio, csn + 1);
+        pio_gpio_init(pio, mosi);
+        pio_sm_set_consecutive_pindirs(pio, sm, csn, 2, true);
+        pio_sm_set_consecutive_pindirs(pio, sm, mosi, 1, true);
+        pio_sm_set_consecutive_pindirs(pio, sm, miso, 1, false);
+
+        pio_sm_config c = sram_reset_program_get_default_config(offset);
+        sm_config_set_in_pins(&c, miso);
+        sm_config_set_in_shift(&c, false, true, 32);
+        sm_config_set_out_pins(&c, mosi, 1);
+        sm_config_set_out_shift(&c, false, true, 32);
+        sm_config_set_sideset_pins(&c, csn);
+        sm_config_set_clkdiv(&c, 4.f);
+
+        pio_sm_init(pio, sm, offset, &c);
+        pio_sm_set_enabled(pio, sm, true);
+    }
+    void aps6404_program_init(PIO pio, uint sm, uint offset, uint csn, uint mosi, bool slow, bool fast, bool reset) {
+        pio_gpio_init(pio, csn);
+        pio_gpio_init(pio, csn + 1);
+        pio_gpio_init(pio, mosi);
+        pio_gpio_init(pio, mosi + 1);
+        pio_gpio_init(pio, mosi + 2);
+        pio_gpio_init(pio, mosi + 3);
+        pio_sm_set_consecutive_pindirs(pio, sm, csn, 2, true);
+        pio_sm_set_consecutive_pindirs(pio, sm, mosi, 4, false);
+
+        pio_sm_config c = slow ? sram_slow_program_get_default_config(offset) : 
+                        fast ? sram_fast_program_get_default_config(offset) : 
+                        reset ? sram_reset_qpi_program_get_default_config(offset) :
+                        sram_program_get_default_config(offset);
+        sm_config_set_in_pins(&c, mosi);
+        sm_config_set_in_shift(&c, false, true, 32);
+        sm_config_set_out_pins(&c, mosi, 4);
+        sm_config_set_out_shift(&c, false, true, 32);
+        sm_config_set_set_pins(&c, mosi, 4);
+        sm_config_set_sideset_pins(&c, csn);
+
+        pio_sm_init(pio, sm, offset, &c);
+        pio_sm_set_enabled(pio, sm, true);
+    }
+}
+
 static const pio_program* pio_prog[2] = {nullptr, nullptr};
 static uint16_t pio_offset[2] = {0xffff, 0xffff};
 
@@ -57,6 +104,7 @@ namespace pimoroni {
         // Claim DMA channels
         dma_channel = dma_claim_unused_channel(true);
         read_cmd_dma_channel = dma_claim_unused_channel(true);
+        setup_dma_config();
     }
 
     void APS6404::init() {
@@ -101,12 +149,13 @@ namespace pimoroni {
 
     void APS6404::adjust_clock() {
         pio_sm_set_enabled(pio, pio_sm, false);
+        uint32_t clock_hz = clock_get_hz(clk_sys);
 
-        if (clock_get_hz(clk_sys) > 296000000) {
+        if (clock_hz > 296000000) {
             pio_offset = pio_change_exclusive_program(pio, &sram_fast_program);
             aps6404_program_init(pio, pio_sm, pio_offset, pin_csn, pin_d0, false, true, false);
         }
-        else if (clock_get_hz(clk_sys) < 130000000) {
+        else if (clock_hz < 130000000) {
             pio_offset = pio_change_exclusive_program(pio, &sram_slow_program);
             aps6404_program_init(pio, pio_sm, pio_offset, pin_csn, pin_d0, true, false, false);
         }
@@ -116,193 +165,16 @@ namespace pimoroni {
         }
 
         last_cmd_was_write = false;
+        page_smashing_ok = clock_hz <= 168000000;
     }
 
-    void APS6404::write(uint32_t addr, uint32_t* data, uint32_t len_in_bytes) {
-        if (!last_cmd_was_write) {
-            dma_channel_config c = dma_channel_get_default_config(dma_channel);
-            channel_config_set_read_increment(&c, true);
-            channel_config_set_write_increment(&c, false);
-            channel_config_set_dreq(&c, pio_get_dreq(pio, pio_sm, true));
-            channel_config_set_transfer_data_size(&c, DMA_SIZE_32);
-            channel_config_set_bswap(&c, true);
-            
-            wait_for_finish_blocking();
-            setup_cmd_buffer_dma(true);
-            last_cmd_was_write = true;
+    
 
-            dma_channel_configure(
-                dma_channel, &c,
-                &pio->txf[pio_sm],
-                nullptr,
-                1,
-                false
-            );
-        }
-
-        int page_len = PAGE_SIZE - (addr & (PAGE_SIZE - 1));
-        int len = len_in_bytes;
-        if ((page_len & 3) != 0) {
-            while (len > page_len) {
-                wait_for_finish_blocking();
-                hw_set_bits(&dma_hw->ch[dma_channel].al1_ctrl, DMA_CH0_CTRL_TRIG_INCR_READ_BITS);
-
-                pio_sm_put_blocking(pio, pio_sm, (page_len << 1) - 1);
-                pio_sm_put_blocking(pio, pio_sm, 0x38000000u | addr);
-                pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_write);
-
-                dma_channel_transfer_from_buffer_now(dma_channel, data, (page_len >> 2) + 1);
-
-                len -= page_len;
-                addr += page_len;
-                data += page_len >> 2;
-                int bytes_sent_last_word = page_len & 3;
-                page_len = std::min(4 - bytes_sent_last_word, len);
-                
-                dma_channel_wait_for_finish_blocking(dma_channel);
-                pio_sm_put_blocking(pio, pio_sm, (page_len << 1) - 1);
-                pio_sm_put_blocking(pio, pio_sm, 0x38000000u | addr);
-                pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_write);
-                pio_sm_put_blocking(pio, pio_sm, __builtin_bswap32(*data >> (8 * bytes_sent_last_word)));
-                
-                addr += page_len;
-                len -= page_len;
-                ++data;
-                page_len = PAGE_SIZE - page_len;
-            }
-        }
-
-        for (page_len = std::min(page_len, len); 
-            len > 0; 
-            addr += page_len, data += page_len >> 2, len -= page_len, page_len = std::min(PAGE_SIZE, len))
-        {
-            wait_for_finish_blocking();
-            hw_set_bits(&dma_hw->ch[dma_channel].al1_ctrl, DMA_CH0_CTRL_TRIG_INCR_READ_BITS);
-
-            pio_sm_put_blocking(pio, pio_sm, (page_len << 1) - 1);
-            pio_sm_put_blocking(pio, pio_sm, 0x38000000u | addr);
-            pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_write);
-
-            dma_channel_transfer_from_buffer_now(dma_channel, data, (page_len >> 2) + 1);
-        }
-    }
-
-    void APS6404::write_repeat(uint32_t addr, uint32_t data, uint32_t len_in_bytes) {
-        if (!last_cmd_was_write) {
-            dma_channel_config c = dma_channel_get_default_config(dma_channel);
-            channel_config_set_read_increment(&c, false);
-            channel_config_set_write_increment(&c, false);
-            channel_config_set_dreq(&c, pio_get_dreq(pio, pio_sm, true));
-            channel_config_set_transfer_data_size(&c, DMA_SIZE_32);
-            channel_config_set_bswap(&c, true);
-            
-            wait_for_finish_blocking();
-            setup_cmd_buffer_dma(true);
-            last_cmd_was_write = true;
-
-            dma_channel_configure(
-                dma_channel, &c,
-                &pio->txf[pio_sm],
-                nullptr,
-                1,
-                false
-            );
-        }
-
-        int first_page_len = PAGE_SIZE - (addr & (PAGE_SIZE - 1));
-        if ((first_page_len & 3) != 0 && (int)len_in_bytes > first_page_len) {
-            wait_for_finish_blocking();
-            hw_clear_bits(&dma_hw->ch[dma_channel].al1_ctrl, DMA_CH0_CTRL_TRIG_INCR_READ_BITS);
-
-            pio_sm_put_blocking(pio, pio_sm, (first_page_len << 1) - 1);
-            pio_sm_put_blocking(pio, pio_sm, 0x38000000u | addr);
-            pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_write);
-
-            dma_channel_transfer_from_buffer_now(dma_channel, &data, (first_page_len >> 2) + 1);
-
-            len_in_bytes -= first_page_len;
-            addr += first_page_len;
-            int bytes_sent_last_word = first_page_len & 3;
-            first_page_len = PAGE_SIZE;
-            data = (data >> (8 * bytes_sent_last_word)) | (data << (32 - (8 * bytes_sent_last_word)));
-        }
-
-        for (int len = len_in_bytes, page_len = std::min(first_page_len, len); 
-             len > 0; 
-             addr += page_len, len -= page_len, page_len = std::min(PAGE_SIZE, len))
-        {
-            wait_for_finish_blocking();
-            hw_clear_bits(&dma_hw->ch[dma_channel].al1_ctrl, DMA_CH0_CTRL_TRIG_INCR_READ_BITS);
-            repeat_data = data;
-
-            pio_sm_put_blocking(pio, pio_sm, (page_len << 1) - 1);
-            pio_sm_put_blocking(pio, pio_sm, 0x38000000u | addr);
-            pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_write);
-
-            dma_channel_transfer_from_buffer_now(dma_channel, &repeat_data, (page_len >> 2) + 1);
-        }
-    }
-
-    void APS6404::read(uint32_t addr, uint32_t* read_buf, uint32_t len_in_words) {
-        start_read(read_buf, len_in_words);
-
-        uint32_t first_page_len = (PAGE_SIZE - (addr & (PAGE_SIZE - 1)));
-
-        if (first_page_len >= len_in_words << 2) {
-            pio_sm_put_blocking(pio, pio_sm, (len_in_words * 8) - 4);
-            pio_sm_put_blocking(pio, pio_sm, 0xeb000000u | addr);
-            pio_sm_put_blocking(pio, pio_sm, pio_offset + sram_offset_do_read);
-        }
-        else {
-            // Must always reset the cmd buffer DMA because writes use the same DREQ
-            setup_cmd_buffer_dma();
-            uint32_t* cmd_buf = add_read_to_cmd_buffer(multi_read_cmd_buffer, addr, len_in_words);
-            dma_channel_transfer_from_buffer_now(read_cmd_dma_channel, multi_read_cmd_buffer, cmd_buf - multi_read_cmd_buffer);
-        }
-    }
-
-    void APS6404::multi_read(uint32_t* addresses, uint32_t* lengths, uint32_t num_reads, uint32_t* read_buf, int chain_channel) {
-        uint32_t total_len = 0;
-        uint32_t* cmd_buf = multi_read_cmd_buffer;
-        for (uint32_t i = 0; i < num_reads; ++i) {
-            total_len += lengths[i];
-            cmd_buf = add_read_to_cmd_buffer(cmd_buf, addresses[i], lengths[i]);
-        }
-
-        start_read(read_buf, total_len, chain_channel);
-        setup_cmd_buffer_dma();
-
-        dma_channel_transfer_from_buffer_now(read_cmd_dma_channel, multi_read_cmd_buffer, cmd_buf - multi_read_cmd_buffer);
-    }
-
-    void APS6404::start_read(uint32_t* read_buf, uint32_t total_len_in_words, int chain_channel) {
-        dma_channel_config c = dma_channel_get_default_config(dma_channel);
-        channel_config_set_read_increment(&c, false);
-        channel_config_set_write_increment(&c, true);
-        channel_config_set_dreq(&c, pio_get_dreq(pio, pio_sm, false));
-        channel_config_set_transfer_data_size(&c, DMA_SIZE_32);
-        channel_config_set_bswap(&c, true);
-        if (chain_channel >= 0) {
-            channel_config_set_chain_to(&c, chain_channel);
-        }
-        
-        last_cmd_was_write = false;
-        wait_for_finish_blocking();
-
-        dma_channel_configure(
-            dma_channel, &c,
-            read_buf,
-            &pio->rxf[pio_sm],
-            total_len_in_words,
-            true
-        );
-    }
-
-    void APS6404::setup_cmd_buffer_dma(bool clear) {
+    void APS6404::setup_dma_config() {
         dma_channel_config c = dma_channel_get_default_config(read_cmd_dma_channel);
         channel_config_set_read_increment(&c, true);
         channel_config_set_write_increment(&c, false);
-        if (!clear) channel_config_set_dreq(&c, pio_get_dreq(pio, pio_sm, true));
+        channel_config_set_dreq(&c, pio_get_dreq(pio, pio_sm, true));
         channel_config_set_transfer_data_size(&c, DMA_SIZE_32);
         
         dma_channel_configure(
@@ -312,36 +184,19 @@ namespace pimoroni {
             0,
             false
         );
-    }
 
-    uint32_t* APS6404::add_read_to_cmd_buffer(uint32_t* cmd_buf, uint32_t addr, uint32_t len_in_words) {
-        int32_t len_remaining = len_in_words << 2;
-        uint32_t len = std::min((PAGE_SIZE - (addr & (PAGE_SIZE - 1))), (uint32_t)len_remaining);
+        write_config = dma_channel_get_default_config(dma_channel);
+        channel_config_set_read_increment(&write_config, true);
+        channel_config_set_write_increment(&write_config, false);
+        channel_config_set_dreq(&write_config, pio_get_dreq(pio, pio_sm, true));
+        channel_config_set_transfer_data_size(&write_config, DMA_SIZE_32);
+        channel_config_set_bswap(&write_config, true);
 
-        while (true) {
-            if (len < 2) {
-                *cmd_buf++ = 0;
-                *cmd_buf++ = 0xeb000000u | addr;
-                *cmd_buf++ = pio_offset + sram_offset_do_read_one;
-            }
-            else {
-                *cmd_buf++ = (len * 2) - 4;
-                *cmd_buf++ = 0xeb000000u | addr;
-                *cmd_buf++ = pio_offset + sram_offset_do_read;
-            }
-            len_remaining -= len;
-            addr += len;
-
-            if (len_remaining <= 0) break;
-
-            len = len_remaining;
-            if (len > PAGE_SIZE) len = PAGE_SIZE;
-        }
-
-        return cmd_buf;
-    }
-
-    void APS6404::wait_for_finish_blocking() {
-        dma_channel_wait_for_finish_blocking(dma_channel);
+        read_config = dma_channel_get_default_config(dma_channel);
+        channel_config_set_read_increment(&read_config, false);
+        channel_config_set_write_increment(&read_config, true);
+        channel_config_set_dreq(&read_config, pio_get_dreq(pio, pio_sm, false));
+        channel_config_set_transfer_data_size(&read_config, DMA_SIZE_32);
+        channel_config_set_bswap(&read_config, true);
     }
 }
