+++ drivers/aps6404/aps6404.pio
@@ -203,51 +203,3 @@ wr_lp:
   out null, 32  [7] side 0b01  ; Ensure large enough gap between commands for reset time
   set pindirs, 0 [7] side 0b01
 .wrap
-
-
-% c-sdk {
-void aps6404_reset_program_init(PIO pio, uint sm, uint offset, uint csn, uint mosi) {
-    uint miso = mosi + 1;
-    pio_gpio_init(pio, csn);
-    pio_gpio_init(pio, csn + 1);
-    pio_gpio_init(pio, mosi);
-    pio_sm_set_consecutive_pindirs(pio, sm, csn, 2, true);
-    pio_sm_set_consecutive_pindirs(pio, sm, mosi, 1, true);
-    pio_sm_set_consecutive_pindirs(pio, sm, miso, 1, false);
-
-    pio_sm_config c = sram_reset_program_get_default_config(offset);
-    sm_config_set_in_pins(&c, miso);
-    sm_config_set_in_shift(&c, false, true, 32);
-    sm_config_set_out_pins(&c, mosi, 1);
-    sm_config_set_out_shift(&c, false, true, 32);
-    sm_config_set_sideset_pins(&c, csn);
-    sm_config_set_clkdiv(&c, 4.f);
-
-    pio_sm_init(pio, sm, offset, &c);
-    pio_sm_set_enabled(pio, sm, true);
-}
-void aps6404_program_init(PIO pio, uint sm, uint offset, uint csn, uint mosi, bool slow, bool fast, bool reset) {
-    pio_gpio_init(pio, csn);
-    pio_gpio_init(pio, csn + 1);
-    pio_gpio_init(pio, mosi);
-    pio_gpio_init(pio, mosi + 1);
-    pio_gpio_init(pio, mosi + 2);
-    pio_gpio_init(pio, mosi + 3);
-    pio_sm_set_consecutive_pindirs(pio, sm, csn, 2, true);
-    pio_sm_set_consecutive_pindirs(pio, sm, mosi, 4, false);
-
-    pio_sm_config c = slow ? sram_slow_program_get_default_config(offset) : 
-                      fast ? sram_fast_program_get_default_config(offset) : 
-                      reset ? sram_reset_qpi_program_get_default_config(offset) :
-                      sram_program_get_default_config(offset);
-    sm_config_set_in_pins(&c, mosi);
-    sm_config_set_in_shift(&c, false, true, 32);
-    sm_config_set_out_pins(&c, mosi, 4);
-    sm_config_set_out_shift(&c, false, true, 32);
-    sm_config_set_set_pins(&c, mosi, 4);
-    sm_config_set_sideset_pins(&c, csn);
-
-    pio_sm_init(pio, sm, offset, &c);
-    pio_sm_set_enabled(pio, sm, true);
-}
-%}
